move connect_to_host, create_firewall_mapping, ip_callback, selected_devices_from_table to Utilities class
